\name{melanoma}
\alias{melanoma}
\title{Melanoma survival data}
\usage{data(melanoma)}
\description{Simulated melanoma survival data.}
\format{
    \tabular{ll}{
	time:\tab survival or censoring time\cr
	status:\tab censoring status\cr
	size: \tab tumor size\cr
	stage: \tab melanoma stage (factor)\cr
	age: \tab patient age (factor)\cr
     }
 }
\keyword{datasets}
