\name{melanoma}
\alias{melanoma}
\title{Melanoma survival data}
\usage{data(melanoma)}
\description{Simulated melanoma survival data.}
\format{
    \tabular{ll}{
	time:\tab survival or censoring time (years)\cr
	status:\tab censoring status (0=censored, 1=death)\cr
	size: \tab tumor size (cm)\cr
	stage: \tab melanoma stage (factor)\cr
	age: \tab patient age (factor)\cr
     }
 }
\keyword{datasets}
